--- a/WalletWasabi/Blockchain/Analysis/Clustering/Cluster.cs
+++ b/WalletWasabi/Blockchain/Analysis/Clustering/Cluster.cs
@@ -9,7 +9,7 @@ namespace WalletWasabi.Blockchain.Analysis.Clustering
 {
 	public class Cluster : NotifyPropertyChangedBase, IEquatable<Cluster>
 	{
-		private List<SmartCoin> Coins { get; set; }
+		public List<SmartCoin> Coins { get; set; }
 
 		private SmartLabel _labels;