--- a/WalletWasabi/Blockchain/TransactionOutputs/CoinLink.cs
+++ b/WalletWasabi/Blockchain/TransactionOutputs/CoinLink.cs
@@ -1,12 +1,4 @@
-using NBitcoin;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Runtime.CompilerServices;
-using WalletWasabi.Bases;
-using WalletWasabi.Helpers;
-using WalletWasabi.Models;
-
+﻿
 namespace WalletWasabi.Blockchain.TransactionOutputs
 {
 	public enum LinkType
@@ -15,12 +7,6 @@ namespace WalletWasabi.Blockchain.TransactionOutputs
 		SpentBy,
 		SameScriptPubKey,
 		SamePubKey
-	}
-
-	public class ILink
-	{
-		SmartCoin Coin { get; }
-		LinkType LinkType { get; }
 	}
 
 	public class CoinLink : ILink