--- a/WalletWasabi/Blockchain/TransactionOutputs/CoinLink.cs
+++ b/WalletWasabi/Blockchain/TransactionOutputs/CoinLink.cs
@@ -1,13 +1,5 @@
 ﻿namespace WalletWasabi.Blockchain.TransactionOutputs
 {
-	public enum LinkType
-	{
-		Spends,
-		SpentBy,
-		SameScriptPubKey,
-		SamePubKey
-	}
-
 	public class CoinLink : ILink
 	{
 		public SmartCoin Coin { get; }