--- a/WalletWasabi/Blockchain/TransactionOutputs/CoinLink.cs
+++ b/WalletWasabi/Blockchain/TransactionOutputs/CoinLink.cs
@@ -2,15 +2,16 @@
 {
 	public class CoinLink : ILink
 	{
-		public SmartCoin Coin { get; }
-		public SmartCoin TargetCoin { get; }
-		public LinkType LinkType { get; }
-
 		public CoinLink(SmartCoin sourceCoin, SmartCoin targetCoin, LinkType linkType)
 		{
 			Coin = sourceCoin;
 			TargetCoin = targetCoin;
 			LinkType = linkType;
 		}
+
+		public SmartCoin Coin { get; }
+		public SmartCoin TargetCoin { get; }
+		public LinkType LinkType { get; }
+
 	}
 }