--- a/WalletWasabi/Blockchain/TransactionOutputs/ILink.cs
+++ b/WalletWasabi/Blockchain/TransactionOutputs/ILink.cs
@@ -1,0 +1,8 @@
+namespace WalletWasabi.Blockchain.TransactionOutputs
+{
+	public interface ILink
+	{
+		SmartCoin Coin { get; }
+		LinkType LinkType { get; }
+	}
+}