--- a/WalletWasabi/Blockchain/TransactionOutputs/LinkType.cs
+++ b/WalletWasabi/Blockchain/TransactionOutputs/LinkType.cs
@@ -1,0 +1,10 @@
+namespace WalletWasabi.Blockchain.TransactionOutputs
+{
+	public enum LinkType
+	{
+		Spends,
+		SpentBy,
+		SameScriptPubKey,
+		SamePubKey
+	}
+}