Creates and keeps updated the links (relationship) between coins, this can be `SpentBy`, `Spends`, `SameScriptPubKey` and `SamePubKey` (not implemented yet). 

It keeps record of which coins were created/destroyed by each transactions in order to speed up the transaction processing.