@lontivero 2020-01-02
Big jump in performance.

## Master branch

```
|  Method |     Job |     Toolchain | TRANSACTIONS |       Mean |     Error |    StdDev | Rank |
|-------- |-------- |-------------- |------------- |-----------:|----------:|----------:|-----:|
| Process | Default | netcoreapp3.0 |          100 |   1.631 ms | 0.0198 ms | 0.0175 ms |    1 |
| Process |    Core |       Default |          100 |   1.718 ms | 0.0328 ms | 0.0438 ms |    2 |
| Process | Default | netcoreapp3.0 |         1000 |  17.340 ms | 0.3432 ms | 0.4462 ms |    4 |
| Process |    Core |       Default |         1000 |  16.800 ms | 0.3315 ms | 0.3256 ms |    3 |
| Process | Default | netcoreapp3.0 |        10000 | 174.926 ms | 3.4253 ms | 5.8164 ms |    5 |
| Process |    Core |       Default |        10000 | 181.490 ms | 3.6190 ms | 5.8440 ms |    6 |
```

## PR

```
|  Method |     Toolchain | TRANSACTIONS |       Mean |     Error |    StdDev | Rank |
|-------- |-------------- |------------- |-----------:|----------:|----------:|-----:|
| Process |       Default |          100 |   1.103 ms | 0.0288 ms | 0.0836 ms |    1 |
| Process | netcoreapp3.1 |          100 |   1.191 ms | 0.0237 ms | 0.0489 ms |    2 |
| Process |       Default |         1000 |  11.242 ms | 0.2246 ms | 0.3992 ms |    3 |
| Process | netcoreapp3.1 |         1000 |  11.253 ms | 0.2227 ms | 0.2083 ms |    3 |
| Process |       Default |        10000 | 111.662 ms | 2.1778 ms | 2.5925 ms |    4 |
| Process | netcoreapp3.1 |        10000 | 118.009 ms | 2.3500 ms | 6.1495 ms |    5 |
```