@nopara73 2020-01-03
With this you'll have to micromanage the smartcoins when you do something with them. IMO that's not good. Please explain the "Big jump in performance" as I don't see it from your table, in fact, I see worse performance at some cases.