@lontivero 2020-01-03
The data table shows that the performance is better in every single scenario. Current version on master requires 181.490 ms to process 10,000 txs while the version in this PR does the same in 118.009 ms (more than a 35% performance gain). It is not possible you see worse performance in any cases, or at least no in the tables that I shared above.

What do you mean by "micromanage the smartcoins"?