@nopara73 2020-01-04
> 35% performance gain

Doesn't seem huge considering the drawback.

> The data table shows that the performance is better in every single scenario.

StdDev for 10000txs on netcore 3.1 went from 5.8440 ms to 6.1495 ms. Maybe I just don't understand how to read the table. Can you explain?

> What do you mean by "micromanage the smartcoins"?

You must keep the `Links` property of smartcoin up to date at all times.