@lontivero 2020-01-06
> Doesn't seem huge considering the drawback.

There is a few algorithms that are critical for performance.This is one of those.

> StdDev for 10000txs on netcore 3.1 went from 5.8440 ms to 6.1495 ms. Maybe I just don't understand how to read the table. Can you explain?

Sure. The standard deviation is a measure of dispersion. Three times the standard deviation cover the 99.7% of the cases. In our case:

* **In master**, the mean (181.490ms) +/- 3 times the stddev (5.8440 ms)  tells us that the 99.7% of the measurements fall in the range of **[163.838ms, 199.14ms]**

* **In this PR**, the mean (118.009ms) +/- 3 times the stddev (6.1495 ms)  tells us that the 99.7% of the measurements fall in the range of **[99.5605ms, 136.4575ms]**

You can see that this PR outperforms the one in master **always** with no exception, even the best time of the current algorithm is worse than the worst time of this algorithm. Even more, in the extreme it can run almost 63% faster.

> You must keep the Links property of smartcoin up to date at all times.

True. That's what I do, this could be a problem if there is a bug and we don't detect it.