@nopara73 2020-01-08
Please create a preparation pull request for this, for example with this "simple" change: https://github.com/zkSNACKs/WalletWasabi/pull/2532/files#diff-98c73da803277a2acbb8c63862ba170e