@nopara73 2020-01-09
> Could you explain what do you expect?

You make that change, and we note, that is not good. In order to do that you have to make the class immutable, which is a fair amount of refactoring. And this is just one "simple" change.