@lontivero 2020-01-10
Let me be more specific. 

* What do you mean by "preparation pull request"?
* What should that other PR contain? What's the difference with this PR? 
* What class should be immutable?

Even more important question: do you think we are going to merge this (or that other PR) one day? Because if you are against the idea for whatever reason then it makes no sense to continue.