@nopara73 2020-01-11
I think the speed increase is impressive, but due to the future maintenance issues with coin link micromanagement I don't want to merge dirty code, thus I suggested to implement it as cleanly as possible. 

You will need to expose `private List<SmartCoin> Coins { get; set; }` to be public as a preparation for this PR. So that's all the first preparation PR should contain, a change of one word: `private`, to `public`. From there on we can talk about why that's good or not good. I just hinted the immutability requirement here, because I don't want to hijack the whole PR talking about that.