@molnard 2020-01-14
Is this a PR to benchmark this? Where are these dictionaries stored? Or are they generated every time when Wasabi started?