@lontivero 2020-01-14
> Is this a PR to benchmark this? 

This PR is for improving the performance and be able to track relationship between coins.

> Where are these dictionaries stored? Or are they generated every time when Wasabi started?

They are generated every time wasabi starts in the same way that the coins are.