@lontivero 2020-02-13
This is amazing, after adapting the @NicolasDorier idea to this PR the gain in performance is crazy:

```
|  Method |     Toolchain | TRANSACTIONS |        Mean |     Error |    StdDev | Rank |
|-------- |-------------- |------------- |------------:|----------:|----------:|-----:|
| Process |       Default |          100 |    116.7 us |   2.32 us |   2.06 us |    1 |
| Process | netcoreapp3.1 |          100 |    121.1 us |   1.48 us |   1.24 us |    2 |
| Process |       Default |         1000 |    979.3 us |  27.57 us |  27.07 us |    4 |
| Process | netcoreapp3.1 |         1000 |    949.4 us |  13.55 us |  12.01 us |    3 |
| Process |       Default |        10000 | 10,345.7 us | 133.85 us | 111.77 us |    5 |
| Process | netcoreapp3.1 |        10000 | 10,661.7 us | 210.58 us | 351.83 us |    5 |
``` 

This is almost 20 times faster.