@lontivero 2020-02-13
In master it takes `181.490 ms` and now it takes `10,661.7 us`. Note the first one is in milliseconds (ms) while the latest one is in microseconds (us). This means: 0.18149s vs 0.0106617s