@nopara73 2020-02-16
@molnard No. It's waiting for dependency: https://github.com/zkSNACKs/WalletWasabi/pull/3080