@lontivero 2020-03-12
The idea behind this PR was to keep track of the coins created and destroyed by each transaction and keep a graph of coins relations. In this way it was possible to speed up the wallet loading time by ~20x, however the wallet loading time was already improved dramatically and then one of the benefits of this PR is no so important.