@yahiheb 2020-03-12
Can we speed up the wallet loading time more with this PR? or it is not worth it considering the trade offs?