@lontivero 2020-03-12
This PR can speed up the wallet loading time, yes. The trade off here is the added complexity.